Updated style.css to account for the new layout
We're moving the builds display from a single 'block' layout to a more modular one, which required the 'Relevant Build' section to be entirely rewritten. There are also some various CSS style fixes here as well. This is the final CSS implementation planned before the main focus will be on build submissions.